Add capability for diff. cell model at diff. nodes
Cell models are now specified as part of specifying the problem in the
"create" type file. Cell models are assigned numerical values, then each
node in the problem is assigned a number to inform which cell model to
use there. Model files are still vectorised, and run for all nodes using
that model simultaneously